[#288] Extract TestClock configuration from SyncJob tests
Until now, the TestClock was only used in SyncJob, but with EUMETSAT
license it is useful to have it extracted.
