+++ s4e-backend/src/main/java/pl/cyfronet/s4e/config/TimeConfig.java
@@ -1,14 +1,23 @@
 package pl.cyfronet.s4e.config;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 import javax.annotation.PostConstruct;
+import java.time.Clock;
 import java.util.TimeZone;
 
 @Configuration
-public class TimeZoneConfig {
+public class TimeConfig {
     @PostConstruct
     void started() {
         TimeZone.setDefault(TimeZone.getTimeZone("UTC"));
+    }
+
+    @Bean
+    @Profile({ "!test", "integration" })
+    public Clock clock() {
+        return Clock.systemUTC();
     }
 }
