+++ s4e-backend/src/test/java/pl/cyfronet/s4e/BasicTest.java
@@ -3,6 +3,7 @@ package pl.cyfronet.s4e;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
 import java.lang.annotation.ElementType;
@@ -12,6 +13,7 @@ import java.lang.annotation.Target;
 
 @SpringBootTest
 @ActiveProfiles("test")
+@Import({ TestClockConfiguration.class, TestMockConfiguration.class })
 @ExtendWith(MockitoExtension.class)
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
