+++ s4e-backend/src/test/java/pl/cyfronet/s4e/TestMockConfiguration.java
@@ -1,16 +1,12 @@
 package pl.cyfronet.s4e;
 
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 import pl.cyfronet.s4e.service.FileStorage;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 
 import static org.mockito.Mockito.mock;
 
-@Configuration
-@Profile("test & !integration")
 public class TestMockConfiguration {
     @Bean
     public FileStorage fileStorage() {
